Forgot to save changes to Vec3...

+++ src/vec3.rs
@@ -1,26 +1,26 @@
 struct Vec3
 {
-    e: [f32; 3],
+    e: [f64; 3],
 }
 
 impl Vec3
 {
-    fn new(x: f32, y: f32, z: f32) -> Vec3
+    fn new(x: f64, y: f64, z: f64) -> Vec3
     {
         Vec3{e: [x,y,z]}
     }
 
-    fn x(&self) -> f32
+    fn x(&self) -> f64
     {
         self.e[0]
     }
 
-    fn y(&self) -> f32
+    fn y(&self) -> f64
     {
         self.e[1]
     }
 
-    fn z(&self) -> f32
+    fn z(&self) -> f64
     {
         self.e[2]
     }
